Simplify the WPF SimpleDemo example. Move Controller code to ControllerDemo in WpfExamples.
