+++ Source/Examples/WPF/SimpleDemo/MainWindow.xaml
@@ -9,7 +9,6 @@
     </Window.DataContext>
     <Grid>
         <!-- The OxyPlot control is binding to a PlotModel in the MainViewModel -->
-        <oxy:PlotView Model="{Binding Model}"
-                      Controller="{Binding Controller}" />
+        <oxy:PlotView Model="{Binding Model}" />
     </Grid>
 </Window>
