+++ Source/Examples/WPF/SimpleDemo/SimpleDemo.csproj
@@ -63,7 +63,6 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="CustomPlotController.cs" />
     <Compile Include="MainViewModel.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
@@ -114,4 +113,3 @@
   </Target>
   -->
 </Project>
-
