+++ Source/Examples/WPF/WpfExamples/Examples/ControllerDemo/MainWindow.xaml
@@ -1,0 +1,15 @@
+﻿<Window x:Class="ControllerDemo.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
+        xmlns:oxy="clr-namespace:OxyPlot.Wpf;assembly=OxyPlot.Wpf" 
+        xmlns:local="clr-namespace:ControllerDemo"
+        Title="OxyPlot Controller Demo" Height="480" Width="640">
+    <Window.DataContext>
+        <local:MainViewModel />
+    </Window.DataContext>
+    <Grid>
+        <!-- The OxyPlot control is binding to a PlotModel and a PlotController in the MainViewModel -->
+        <oxy:PlotView Model="{Binding Model}"
+                      Controller="{Binding Controller}" />
+    </Grid>
+</Window>
