+++ Source/Examples/WPF/WpfExamples/WpfExamples.csproj
@@ -69,6 +69,11 @@
     <Compile Include="Examples\AlignedAxesDemo\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Examples\ControllerDemo\CustomPlotController.cs" />
+    <Compile Include="Examples\ControllerDemo\MainViewModel.cs" />
+    <Compile Include="Examples\ControllerDemo\MainWindow.xaml.cs">
+      <DependentUpon>MainWindow.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Examples\LinearBarSeriesDemo\MainViewModel.cs" />
     <Compile Include="Examples\LinearBarSeriesDemo\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
@@ -98,6 +103,10 @@
     </Compile>
     <Compile Include="ScreenCapture.cs" />
     <Page Include="Examples\AnnotationDemo\MainWindow.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Examples\ControllerDemo\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -415,6 +424,7 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
+    <Resource Include="Images\ControllerDemo.png" />
     <Resource Include="Images\ToolTipDemo.png" />
     <Resource Include="Images\LinearBarSeriesDemo.png" />
     <Resource Include="Images\DataTemplateDemo.png" />
@@ -504,4 +514,3 @@
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
-
